blog_engine/pyproject.toml
[project]
name = "blog-engine"
version = "0.1.0"
description = "B-zone server for content generation, SEO, WordPress publishing"
authors = [{ name = "seungminsong", email = "dev@example.com" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dynamic = ["dependencies"]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.116.1"
uvicorn = { version = "0.35.0", extras = ["standard"] }
SQLAlchemy = "2.0.43"
pymysql = "1.1.1"
cryptography = "^46.0.3"
alembic = "1.16.5"
pydantic-settings = "2.10.1"
openai = "1.102.0"
requests = "2.32.5"
Pillow = "11.3.0"
Jinja2 = "3.1.6"
Markdown = "3.8.2"
python-slugify = "8.0.4"
python-multipart = "0.0.20"
google-api-python-client = "2.179.0"
google-auth = "2.40.3"
slowapi = "0.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
